babel-plugin-react-intl as preset not plugin
Babel executes plugins from the highest level first, then
executes plugins inside presets.

By pushing `babel-plugin-react-intl` into `plugins` at the last
moment it meant it always ran first which doesn't allow for the user
to run any babel presets BEFORE `babel-plugin-react-intl` which can cause
issues.

This changes it to be included as a plugin within a preset, and makes that
the LAST preset to be executed (right to left), thereby making it the lowest
priority plugin operation.